+++ src/modules/widget/base/button/Schema.js
@@ -1,7 +1,7 @@
 
-import Schema from '../../../schema/Schema'
+import { BaseSchema } from '../../../schema'
 
-export default class ButtonSchema extends Schema {
+export default class ButtonSchema extends BaseSchema {
   constructor (props) {
     super()
     this.option = {