+++ src/modules/widget/base/formButton/Schema.js
@@ -1,6 +1,6 @@
-import Schema from '../../../schema/Schema'
+import { BaseSchema } from '../../../schema'
 
-export default class ButtonSchema extends Schema {
+export default class ButtonSchema extends BaseSchema {
   constructor (props) {
     super()
     this.option = {