+++ src/modules/widget/base/rate/Schema.js
@@ -1,6 +1,6 @@
-import Schema from '../../../schema/FormSchema'
+import { FormSchema } from '../../../schema'
 
-export default class RateSchema extends Schema {
+export default class RateSchema extends FormSchema {
   constructor (props) {
     super()
     this.type = 'number'